Atualizando classe Pet

+++ backend/src/main/java/com/github/laertty/sistema_cadastro_pets/repository/PetRepository.java
@@ -1,8 +1,7 @@
 package com.github.laertty.sistema_cadastro_pets.repository;
 
-
 import com.github.laertty.sistema_cadastro_pets.model.entities.Pet;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PetRepository extends CrudRepository<Pet, Long> {}
+public interface PetRepository extends CrudRepository<Pet, Integer> {}
